requirements.txt
azure-identity==1.13.0
azure-keyvault-secrets==4.7.0
